+++ examples/utils/__init__.py
@@ -1,27 +1,1 @@
 """Utility functions for BeCoMe examples."""
-
-# Re-export all public functions for backward compatibility
-from .analysis import calculate_agreement_level
-from .data_loading import load_data_from_txt
-from .display import (
-    display_median_calculation_details,
-    display_step_1_arithmetic_mean,
-    display_step_2_median,
-    display_step_3_best_compromise,
-    display_step_4_max_error,
-)
-from .formatting import display_case_header, display_centroid, print_header, print_section
-
-__all__ = [
-    "load_data_from_txt",
-    "print_header",
-    "print_section",
-    "display_case_header",
-    "display_centroid",
-    "display_step_1_arithmetic_mean",
-    "display_step_2_median",
-    "display_step_3_best_compromise",
-    "display_step_4_max_error",
-    "display_median_calculation_details",
-    "calculate_agreement_level",
-]
