+++ src/__init__.py
@@ -1,17 +1,3 @@
 """BeCoMe (Best Compromise Mean) method for group decision-making."""
 
-from .exceptions import (
-    BeCoMeError,
-    CalculationError,
-    EmptyOpinionsError,
-    InvalidOpinionError,
-)
-
 __version__ = "0.1.0"
-
-__all__ = [
-    "BeCoMeError",
-    "CalculationError",
-    "EmptyOpinionsError",
-    "InvalidOpinionError",
-]
