+++ src/interpreters/__init__.py
@@ -1,8 +1,1 @@
 """Interpreters for BeCoMe results."""
-
-from .likert_interpreter import LikertDecision, LikertDecisionInterpreter
-
-__all__ = [
-    "LikertDecision",
-    "LikertDecisionInterpreter",
-]
