+++ src/models/__init__.py
@@ -1,7 +1,1 @@
 """BeCoMe data models."""
-
-from .become_result import BeCoMeResult
-from .expert_opinion import ExpertOpinion
-from .fuzzy_number import FuzzyTriangleNumber
-
-__all__ = ["BeCoMeResult", "ExpertOpinion", "FuzzyTriangleNumber"]
