+++ tests/reference/__init__.py
@@ -1,6 +1,6 @@
 """Reference test cases from Excel implementation.
 
-This module exports three case studies used for validation against Excel calculations:
+This module contains three case studies used for validation against Excel calculations:
 
 BUDGET_CASE:
     22 experts (even), interval estimates for COVID-19 pandemic budget support
@@ -14,9 +14,3 @@ FLOODS_CASE:
 Each case contains expert opinions loaded from examples/data/ and expected results
 from Excel calculations for verification of BeCoMe algorithm implementation.
 """
-
-from tests.reference.budget_case import BUDGET_CASE
-from tests.reference.floods_case import FLOODS_CASE
-from tests.reference.pendlers_case import PENDLERS_CASE
-
-__all__ = ["BUDGET_CASE", "FLOODS_CASE", "PENDLERS_CASE"]
